+++ src/data/changelog-version.ts
@@ -1,3 +1,3 @@
 export function getLatestVersion(): { version: string; date: string } {
-  return { version: "8.95.4", date: "2026-03-21" };
+  return { version: "8.95.5", date: "2026-03-21" };
 }